Extract logging to own file
